Update test section
- update Quick&Nimble templates
- add custom Quick&Nimble template
- update XCTests templates
- add custom XCTests template
