+++ Templates/File Template/Tests/QuickSpecPresenterTests.xctemplate/TemplateIcon.png
