+++ Templates/File Template/Tests/QuickSpecPresenterTests.xctemplate/TemplateIcon@2x.png
