+++ Templates/File Template/Tests/QuickSpecPresenterTests.xctemplate/TemplateInfo.plist
@@ -5,14 +5,14 @@
 	<key>DefaultCompletionName</key>
 	<string>MVVM QuickSpec</string>
 	<key>Description</key>
-	<string>Contains Quick framework init Test template for project with MVVM usage</string>
+	<string>Contains Quick framework init Test template for project with VIPER or VIP usage</string>
 	<key>Kind</key>
 	<string>Xcode.IDEKit.TextSubstitutionFileTemplateKind</string>
 	<key>Options</key>
 	<array>
 		<dict>
 			<key>Description</key>
-			<string>The name of the View Model to test</string>
+			<string>The name of the Prenter to test</string>
 			<key>Identifier</key>
 			<string>sceneName</string>
 			<key>Name</key>
@@ -34,13 +34,13 @@
 		</dict>
 		<dict>
 			<key>Default</key>
-			<string>___VARIABLE_sceneName:identifier___ViewModelTests</string>
+			<string>___VARIABLE_sceneName:identifier___PresenterTests</string>
 			<key>Description</key>
-			<string>View Model Unit Test name</string>
+			<string>Presenter Unit Test name</string>
 			<key>Identifier</key>
-			<string>viewModel</string>
+			<string>presenter</string>
 			<key>Name</key>
-			<string>View Model Test Name:</string>
+			<string>Presenter Test Name:</string>
 			<key>Required</key>
 			<true/>
 			<key>Type</key>
@@ -50,7 +50,7 @@
 	<key>SortOrder</key>
 	<string>9</string>
 	<key>Summary</key>
-	<string>Simple QuickSpec test for VM</string>
+	<string>Simple QuickSpec test for Presenter</string>
 	<key>SupportsSwiftPackage</key>
 	<true/>
 </dict>
