+++ Templates/File Template/Tests/QuickSpecPresenterTests.xctemplate/___FILEBASENAME__PresenterTests.swift
@@ -1,0 +1,21 @@
+import Quick
+import Nimble
+@testable import ___PROJECTNAME___
+
+// MARK: - ___VARIABLE_sceneName___PresenterTests
+final class ___VARIABLE_sceneName___PresenterTests: QuickSpec {
+    override class func spec() {
+        // Mocks 
+
+        // System under tests
+        var sut: ___VARIABLE_sceneName___Presenter!
+
+        beforeEach {
+            sut = .init()
+        }
+
+        describe("") {
+            context("") {}
+        }
+    }
+}
