+++ Templates/File Template/Tests/QuickSpecTests.xctemplate/TemplateIcon.png
