+++ Templates/File Template/Tests/QuickSpecTests.xctemplate/TemplateIcon@2x.png
