+++ Templates/File Template/Tests/QuickSpecTests.xctemplate/___FILEBASENAME___Tests.swift
@@ -1,0 +1,21 @@
+import Quick
+import Nimble
+@testable import ___PROJECTNAME___
+
+// MARK: - ___VARIABLE_sceneName___Tests
+final class ___VARIABLE_sceneName___Tests: QuickSpec {
+    override class func spec() {
+        // Mocks 
+
+        // System under tests
+        // var sut: YourType
+
+        beforeEach {
+            sut = .init()
+        }
+
+        describe("") {
+            context("") {}
+        }
+    }
+}
